avoid forcing inlining for amd processors
