+++ src/object.c
@@ -328,33 +328,6 @@ kvobj *kvobjSet(sds key, robj *val, uint32_t keyMetaBits) {
     return kv;
 }
 
-/* Create a string object with EMBSTR encoding if it is smaller than
- * OBJ_ENCODING_EMBSTR_SIZE_LIMIT, otherwise the RAW encoding is
- * used.
- *
- * The current limit of 44 is chosen so that the biggest string object
- * we allocate as EMBSTR will still fit into the 64 byte arena of jemalloc. */
-#define OBJ_ENCODING_EMBSTR_SIZE_LIMIT 44
-robj *createStringObject(const char *ptr, size_t len) {
-    if (len <= OBJ_ENCODING_EMBSTR_SIZE_LIMIT)
-        return createEmbeddedStringObject(ptr,len);
-    else
-        return createRawStringObject(ptr,len);
-}
-
-/* Inline version of createStringObject(), used on hot paths where this call
- * has been measured to be faster than original one.
- * always_inline: force expansion at every call site.
- * flatten:       transitively inline the callees (createEmbeddedStringObject /
- *                createRawStringObject and their helpers) into the caller. */
-__attribute__((always_inline, flatten))
-inline robj *createStringObjectInline(const char *ptr, size_t len) {
-    if (len <= OBJ_ENCODING_EMBSTR_SIZE_LIMIT)
-        return createEmbeddedStringObject(ptr,len);
-    else
-        return createRawStringObject(ptr,len);
-}
-
 /* Same as CreateRawStringObject, can return NULL if allocation fails */
 robj *tryCreateRawStringObject(const char *ptr, size_t len) {
     sds str = sdstrynewlen(ptr,len);
